Fix error when old JDKs are still in path
When an old JDK is listed, but missing from the actual file system, the `jdk` variable in the `forEach` loop will be undefined, causing the error "cannot read property 'version' of undefined" to be thrown. This fix filters out invalid JDKs before they even get to the loop, and still allows the `length` check to do its thing if after the filtering there are no valid JDKs.